Control the build of ethminer using environment variables
Allows the version of ethminer that is built to be determined using
environment variables whose values can be specified in the automine_config json
file.

The following variables are used:

- ETHMINER_GITHUB=<github user name of the ethminer fork to build from>
- ETHMINER_BRANCH=<branch on github to checkout>
- ETHMINER_COMMIT=<commit on github to build on>

All these environment variables are optional, but when set they allow control
over what version of ethminer is built on a rig.